requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
torch>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0
